refactor: remove obsolete configurations and dependencies
