+++ pyproject.toml
@@ -20,7 +20,6 @@ fastapi = {extras = ["standard"], version = "==0.116.1"}
 uvicorn = "==0.35.0"
 # ORM & Database
 advanced-alchemy = "==1.6.2"
-sqlmodel = "==0.0.24"
 psycopg = {extras = ["binary", "pool"], version = "==3.2.10"}
 asyncpg = "==0.30.0"
 aiosqlite = "==0.21.0"
@@ -35,7 +34,7 @@ loguru = "==0.7.3"
 # Application Configuration
 pyhocon = "==0.3.61"
 # Statistics & Data
-numpy = "==2.3.2"
+numpy = "==2.3.3"
 pandas = "==2.3.2"
 # Datetime
 arrow = "==1.3.0"
